Refactor Mastodon client into modules
